weight (Tensor, optional) – a manual rescaling weight given to each class. If given, has to be a Tensor of size C and floating point dtype

+++ presto/eval.py
@@ -352,7 +352,7 @@ class WorldCerealEval:
             target_function=self.target_function,
         )
 
-        loss_fn = nn.CrossEntropyLoss(weight=torch.from_numpy(train_ds.class_weights))
+        loss_fn = nn.CrossEntropyLoss(weight=torch.from_numpy(train_ds.class_weights).float())
 
         generator = torch.Generator()
         generator.manual_seed(self.seed)
